Refactoring common diagnostics method from DataContextV3 to DataContext; also using to_json_dict in Configuration.
